+++ dags/veda_data_pipeline/veda_dataset_pipeline.py
@@ -1,10 +1,16 @@
 import pendulum
+from datetime import timedelta
+
 from airflow import DAG
 from airflow.decorators import task
 from airflow.operators.dummy_operator import DummyOperator as EmptyOperator
 from airflow.utils.trigger_rule import TriggerRule
+from airflow.models.variable import Variable
+from airflow.providers.amazon.aws.operators.ecs import EcsRunTaskOperator
+
 from veda_data_pipeline.groups.collection_group import collection_task_group
-from veda_data_pipeline.groups.discover_group import subdag_discover
+from veda_data_pipeline.groups.discover_group import discover_from_s3_task, get_files_to_process
+from veda_data_pipeline.groups.processing_tasks import build_stac_kwargs, submit_to_stac_ingestor_task
 
 dag_doc_md = """
 ### Dataset Pipeline
@@ -70,11 +76,31 @@ template_dag_run_conf = {
 }
 
 with DAG("veda_dataset_pipeline", params=template_dag_run_conf, **dag_args) as dag:
+    # ECS dependency variable
+    mwaa_stack_conf = Variable.get("MWAA_STACK_CONF", deserialize_json=True)
+
     start = EmptyOperator(task_id="start", dag=dag)
     end = EmptyOperator(task_id="end", dag=dag)
 
     collection_grp = collection_task_group()
-    discover_grp = subdag_discover.expand(event=extract_discovery_items())
+    discover = discover_from_s3_task.expand(event=extract_discovery_items())
+    discover.set_upstream(collection_grp) # do not discover until collection exists
+    get_files = get_files_to_process(payload=discover)
+    build_stac_kwargs_task = build_stac_kwargs.expand(event=get_files)
+    # partial() is needed for the operator to be used with taskflow inputs
+    build_stac = EcsRunTaskOperator.partial(
+        task_id="build_stac",
+        execution_timeout=timedelta(minutes=60),
+        trigger_rule=TriggerRule.NONE_FAILED,
+        cluster=f"{mwaa_stack_conf.get('PREFIX')}-cluster",
+        task_definition=f"{mwaa_stack_conf.get('PREFIX')}-tasks",
+        launch_type="FARGATE",
+        do_xcom_push=True
+    ).expand_kwargs(build_stac_kwargs_task)
+    # .output is needed coming from a non-taskflow operator
+    submit_stac = submit_to_stac_ingestor_task.expand(built_stac=build_stac.output)
 
-    start >> collection_grp >> discover_grp >> end
+    collection_grp.set_upstream(start)
+    submit_stac.set_downstream(end)
+
