+++ dags/veda_data_pipeline/veda_discover_pipeline.py
@@ -1,8 +1,15 @@
 import pendulum
+
+from datetime import timedelta
 from airflow import DAG
 from airflow.operators.dummy_operator import DummyOperator
 from airflow.utils.trigger_rule import TriggerRule
-from veda_data_pipeline.groups.discover_group import subdag_discover
+from airflow.models.variable import Variable
+from airflow.providers.amazon.aws.operators.ecs import EcsRunTaskOperator
+
+from veda_data_pipeline.groups.discover_group import discover_from_s3_task, get_files_to_process
+from veda_data_pipeline.groups.processing_tasks import build_stac_kwargs, submit_to_stac_ingestor_task
+
 
 dag_doc_md = """
 ### Discover files from S3
@@ -46,7 +53,7 @@ dag_args = {
     "is_paused_upon_creation": False,
 }
 
-templat_dag_run_conf = {
+template_dag_run_conf = {
     "collection": "<coll_name>",
     "bucket": "<bucket>",
     "prefix": "<prefix>/",
@@ -70,23 +77,44 @@ templat_dag_run_conf = {
 
 
 def get_discover_dag(id, event={}):
-    params_dag_run_conf = event or templat_dag_run_conf
+    params_dag_run_conf = event or template_dag_run_conf
     with DAG(
         id,
         schedule_interval=event.get("schedule"),
         params=params_dag_run_conf,
         **dag_args
     ) as dag:
+        # ECS dependency variable
+        mwaa_stack_conf = Variable.get("MWAA_STACK_CONF", deserialize_json=True)
+
         start = DummyOperator(task_id="Start", dag=dag)
         end = DummyOperator(
             task_id="End", trigger_rule=TriggerRule.ONE_SUCCESS, dag=dag
         )
+        # define DAG using taskflow notation
+        
+        discover = discover_from_s3_task(event=event)
+        get_files = get_files_to_process(payload=discover)
+        build_stac_kwargs_task = build_stac_kwargs.expand(event=get_files)
+        # partial() is needed for the operator to be used with taskflow inputs
+        build_stac = EcsRunTaskOperator.partial(
+            task_id="build_stac",
+            execution_timeout=timedelta(minutes=60),
+            trigger_rule=TriggerRule.NONE_FAILED,
+            cluster=f"{mwaa_stack_conf.get('PREFIX')}-cluster",
+            task_definition=f"{mwaa_stack_conf.get('PREFIX')}-tasks",
+            launch_type="FARGATE",
+            do_xcom_push=True
+        ).expand_kwargs(build_stac_kwargs_task)
+        # .output is needed coming from a non-taskflow operator
+        submit_stac = submit_to_stac_ingestor_task.expand(built_stac=build_stac.output)
 
-        discover_grp = subdag_discover(event)
-
-        start >> discover_grp >> end
+        discover.set_upstream(start)
+        submit_stac.set_downstream(end)
 
         return dag
+    
+    a >> b >> c
 
 
 get_discover_dag("veda_discover")
